requirements.txt
astroid==2.1.0
autopep8==1.4.3
Click==7.0
Flask==1.0.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
Pillow==5.4.1
pycodestyle==2.5.0
pylint==2.2.2
six==1.12.0
typed-ast==1.3.1
Werkzeug==0.14.1
wrapt==1.11.1
